requirements.txt
boto3
kafka-python
cottonformation